chore: Ajout des TODO pour réaliser les premiers tests
