+++ front/src/app/features/sessions/services/session-api.service.spec.ts
@@ -19,4 +19,29 @@ describe('SessionsService', () => {
   it('should be created', () => {
     expect(service).toBeTruthy();
   });
+
+  //TODO: Affichage de la liste des sessions
+  it('should display the list of sessions', () => {
+
+  });
+
+  //TODO: Les informations de la session sont correctement affichées
+  it('should correctly display session information', () => {
+
+  });
+
+  //TODO: La session est crée
+  it('should create a session', () => {
+
+  });
+
+  //TODO: La session est correctement supprimé
+  it('should correctly delete a session', () => {
+
+  });
+
+  //TODO: La session est modifiée
+  it('should modify a session', () => {
+
+  });
 });
